Do it for real

+++ src/core/DynaRec_x64/recompiler.cc
@@ -279,7 +279,7 @@ void DynaRecCPU::handleLinking() {
             *(uint32_t*)(pointer - 4) = (uint32_t)(uintptr_t)*nextBlockPointer; // Patch comparison value
         } else { // If it has already been compiled, link by jumping to the compiled code
             loadAddress(rax, nextBlockPointer);
-            gen.cmp(dword[rax], (uint32_t)*nextBlockPointer);
+            gen.cmp(dword[rax], (uint32_t)(uintptr_t)*nextBlockPointer);
             gen.jne((void*)m_returnFromBlock); // Return if the block addr changed
             gen.jmp((void*)*nextBlockPointer); // Jump to linked block otherwise
         }
